wip: seprate elliptic data to separate file
